+++ Test/WebDriver/Commands/Internal.hs
@@ -31,7 +31,7 @@ doWinCommand :: (ToJSON a, FromJSON b) =>
 doWinCommand = doWinCommand' []
 
 doStorageCommand :: (ToJSON a, FromJSON b) =>
-                    RequestMethod -> HTML5StorageType -> Text -> a -> WD b
+                    RequestMethod -> WebStorageType -> Text -> a -> WD b
 doStorageCommand = doStorageCommand' []
 
 doCommand' :: (ToJSON a, FromJSON b) => 
@@ -67,7 +67,7 @@ doElemCommand' h m (Element e) path a =
   doSessCommand' h m (T.concat ["/element/", e, path]) a
 
 doStorageCommand' :: (ToJSON a, FromJSON b) =>
-                     [Header] -> RequestMethod -> HTML5StorageType -> Text -> a
+                     [Header] -> RequestMethod -> WebStorageType -> Text -> a
                      -> WD b
 doStorageCommand' h m s path a = doSessCommand' h m (T.concat ["/", s', path]) a
   where s' = case s of